feat: prevent car removal if it is associated with a ride

+++ src/main/java/br/ifsp/demo/usecase/car/RemoveCarUseCase.java
@@ -1,20 +1,27 @@
 package br.ifsp.demo.usecase.car;
 
 import br.ifsp.demo.domain.Driver;
+import br.ifsp.demo.domain.Ride;
+import br.ifsp.demo.exception.CarInUseException;
 import br.ifsp.demo.exception.CarNotFoundException;
 import br.ifsp.demo.exception.DriverNotFoundException;
 import br.ifsp.demo.repositories.CarRepository;
 import br.ifsp.demo.repositories.DriverRepository;
+import br.ifsp.demo.repositories.RideRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.UUID;
 
 @Service
 @AllArgsConstructor
 public class RemoveCarUseCase {
+
     private DriverRepository driverRepository;
     private CarRepository carRepository;
+    private RideRepository rideRepository;
+
     public void execute(UUID driverId, UUID carId) {
         Driver driver = driverRepository.findById(driverId)
                 .orElseThrow(() -> new DriverNotFoundException(driverId));
@@ -26,8 +33,17 @@ public class RemoveCarUseCase {
             throw new CarNotFoundException(carId);
         }
 
+        boolean carIsInUse = rideRepository.findRideByDriver_Id(driverId).stream()
+                .anyMatch(ride -> ride.getCar() != null && ride.getCar().getId().equals(carId));
+
+        if (carIsInUse) {
+            throw new CarInUseException(carId);
+        }
+
         driver.removeCarById(carId);
         driverRepository.save(driver);
         carRepository.deleteById(carId);
     }
+
+
 }
